move the native function/wrapper generation to its own module
